@@ -1,13 +1,20 @@
-import { Component, Input, OnInit, SimpleChanges, ViewChild } from '@angular/core';
+import {
+  Component,
+  Input,
+  OnInit,
+  SimpleChanges,
+  ViewChild,
+} from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
-import { take } from 'rxjs/operators';
-import { AddEditRecipyComponent } from 'src/app/recipies/components/add-edit-recipy/add-edit-recipy.component';
+import { Store } from '@ngrx/store';
 import { Recipy } from 'src/app/recipies/models/recipy.interface';
 import { RecipiesService } from 'src/app/recipies/services/recipies.service';
+
+import * as RecipyActions from '../../../store/actions/recipies.actions';
 
 @Component({
   selector: 'app-recipies-admin',
@@ -18,16 +25,19 @@ export class RecipiesAdminComponent implements OnInit {
   @Input()
   recipies!: Recipy[];
 
-  displayedColumns: string[] = [
-    'name',
-    'complexity',
-    'type',
-    'author',
-    'createdOn',
-    'editedBy',
-    'lastEdited',
-    'isClone',
-    'actions',
+  displayedColumns: string[] = [];
+
+  allColumns = [
+    { name: 'name', displayed: true, selectDisabled: true },
+    { name: 'complexity', displayed: true, selectDisabled: false },
+    { name: 'type', displayed: true, selectDisabled: false },
+    { name: 'author', displayed: false, selectDisabled: false },
+    { name: 'createdOn', displayed: true, selectDisabled: false },
+    { name: 'editedBy', displayed: false, selectDisabled: false },
+    { name: 'lastEdited', displayed: false, selectDisabled: false },
+    { name: 'isClone', displayed: false, selectDisabled: false },
+    { name: 'isCheckedAndApproved', displayed: true, selectDisabled: false },
+    { name: 'actions', displayed: true, selectDisabled: true },
   ];
 
   dataSource!: MatTableDataSource<Recipy>;
@@ -41,7 +51,8 @@ export class RecipiesAdminComponent implements OnInit {
     public dialog: MatDialog,
     private recipiesService: RecipiesService,
     private router: Router,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    private store: Store
   ) {}
   ngOnChanges(changes: SimpleChanges): void {
     this.initTable();
@@ -49,6 +60,9 @@ export class RecipiesAdminComponent implements OnInit {
 
   ngOnInit() {
     this.initTable();
+    this.displayedColumns = this.allColumns
+      .filter((column) => column.displayed)
+      .map((column) => column.name);
   }
 
   initTable(): void {
@@ -71,26 +85,9 @@ export class RecipiesAdminComponent implements OnInit {
   }
 
   addRecipy() {
-    const dialogRef = this.dialog.open(AddEditRecipyComponent, {
-      width: '90%',
-      maxWidth: '100%',
-      height: '80%',
-      hasBackdrop: false,
-      panelClass: 'add-recipy-dialog',
-      autoFocus: false,
-      data: {
-        mode: 'create',
-      },
+    this.router.navigate(['recipies', 'edit-recipy'], {
+      relativeTo: this.route,
     });
-
-    dialogRef
-      .afterClosed()
-      .pipe(take(1))
-      .subscribe((result: any) => {
-        if (result) {
-          this.recipiesService.addRecipy(result.recipy);
-        }
-      });
   }
 
   openRecipy(recipy: Recipy) {
@@ -99,43 +96,51 @@ export class RecipiesAdminComponent implements OnInit {
     });
   }
 
-  editRecipy(recipy: Recipy) {
-    const dialogRef = this.dialog.open(AddEditRecipyComponent, {
-      width: '90%',
-      maxWidth: '100%',
-      height: '80%',
-      hasBackdrop: false,
-      panelClass: 'add-recipy-dialog',
-      autoFocus: false,
-      data: {
-        mode: 'edit',
-        recipy: recipy,
-      },
-    });
-
-    dialogRef
-      .afterClosed()
-      .pipe(take(1))
-      .subscribe((result: any) => {
-        if (result?.recipy.id) {
-          this.recipiesService.updateRecipy(result.recipy);
-        }
-      });
-  }
-
   deleteRecipy(recipy: Recipy) {
     this.recipiesService.deleteRecipy(recipy);
   }
 
-  getAuthor(recipy: Recipy){
-    return this.recipiesService.getRecipyBelongsTo(recipy)
+  getAuthor(recipy: Recipy) {
+    return this.recipiesService.getRecipyBelongsTo(recipy);
   }
 
-  getCreatedOn(recipy: Recipy){
-    return this.recipiesService.getRecipyCreatedOn(recipy)
+  getCreatedOn(recipy: Recipy) {
+    return this.recipiesService.getRecipyCreatedOn(recipy);
   }
 
-  getIsClone(recipy: Recipy): boolean{
-    return Boolean(recipy.clonedBy)
+  getIsClone(recipy: Recipy): boolean {
+    return Boolean(recipy.clonedBy);
+  }
+
+  getIsIngredientNotInDB(recipy: Recipy): boolean {
+    return !!recipy.notApproved;
+  }
+
+  onselectionChange(event: any) {
+    console.log(this.displayedColumns);
+  }
+  onisCheckedAndApprovedClicked(recipy: Recipy) {
+    let updatedRecipy: Recipy = {
+      ...recipy,
+      isCheckedAndApproved: !recipy.isCheckedAndApproved,
+    };
+    this.store.dispatch(new RecipyActions.UpdateRecipyAction(updatedRecipy));
+  }
+
+  updateRecipies() {
+    this.recursiveUpdate(0);
+  }
+  recursiveUpdate(i: number) {
+    if (i < this.recipies.length - 1) {
+      setTimeout(() => {
+        if (!this.recipies[i].calorificValue) {
+          this.store.dispatch(
+            new RecipyActions.UpdateRecipyAction(this.recipies[i])
+          );
+        }
+        i++;
+        this.recursiveUpdate(i);
+      }, 1000);
+    }
   }
 }
