@@ -1,5 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+import { User } from '../models/user.interface';
 
 @Injectable({
   providedIn: 'root'
@@ -16,7 +18,7 @@ export class AuthApiService {
     return this.http.get(`${this.url}.json`)
   }
 
-  updateUser(id: string, data: any){
-    return this.http.patch(`${this.url}/${id}.json`, data)
+  updateUser(id: string, data: any): Observable<User>{
+    return this.http.patch<User>(`${this.url}/${id}.json`, data)
   }
 }
