@@ -1,7 +1,9 @@
 import { Injectable } from '@angular/core';
+import { Store } from '@ngrx/store';
 import { createUserWithEmailAndPassword, getAuth, signInWithEmailAndPassword, signOut } from 'firebase/auth';
 import { BehaviorSubject } from 'rxjs';
 
+import { Role, User } from '../models/user.interface';
 import { UserService } from './user.service';
 
 @Injectable({
@@ -9,8 +11,9 @@ import { UserService } from './user.service';
 })
 export class AuthService {
   isLoggedIn = new BehaviorSubject<boolean>(false);
+  isCheckPerformed$ = new BehaviorSubject<boolean>(false);
 
-  constructor(private userService: UserService) {}
+  constructor(private userService: UserService, private store: Store) {}
 
   registerUser(email: string, password: string) {
     const auth = getAuth();
@@ -55,16 +58,23 @@ export class AuthService {
 
   checkIsLoggedIn() {
     getAuth().onAuthStateChanged((user) => {
+      this.isCheckPerformed$.next(true);
       if (user) {
+        if (user.email) {
+          let currentUser: User = {
+            email: user.email,
+            uid: user.uid,
+            role: Role.User
+          };
+          // this.store.dispatch(new UserActions.UserLoadedAction(currentUser))
+        }
+
         this.userService.userAtFirebaseAuth = user;
+        console.log(user);
         this.isLoggedIn.next(true);
       } else {
         this.isLoggedIn.next(false);
       }
     });
   }
-
-
-
-
 }
