@@ -11,5 +11,16 @@ export enum DishType {
   'дитяче меню' = 10,
   'дієтичне меню' = 11,
   'вегетаріанські страви' = 12,
-  'сніданок'
+  'сніданок' = 13,
+  'потребує попередньої підготовки' = 14,
+  'м\'ясна страва' = 15,
+  'гостра страва' = 16,
+  'приготування в духовці' = 17,
+  'на грилі' = 18,
+  'мультиварка' = 19,
+  'на пару' = 20,
+  'варіння' = 21,
+  'без термообробки' = 22,
+  'гарнір' = 23,
+  'риба і морепродукти' = 24
 }
