@@ -1,9 +1,9 @@
 import { Complexity } from './complexity.enum';
 import { DishType } from './dishType.enum';
-import { Ingredient } from './ingredient.interface';
+import { Ingredient, IngredientsGroup } from './ingredient.interface';
 import { PreparationStep } from './preparationStep.interface';
 
-export interface Recipy {
+export interface Recipy extends NewRecipy {
   id: string;
   name: string;
   ingrediends: Ingredient[];
@@ -17,7 +17,16 @@ export interface Recipy {
   lastEdited?: number;
   clonedBy?: string;
   clonedOn?: number;
-  originalRecipy?: string
+  originalRecipy?: string;
+  isSplitIntoGroups: IngredientsGroup[];
+  isCheckedAndApproved?: boolean;
+  calorificValue?: number;
+  source: string;
+}
+
+export interface RecipyForCalendar extends Recipy {
+  portions: number,
+  amountPerPortion: number
 }
 
 export interface NewRecipy {
@@ -31,5 +40,26 @@ export interface NewRecipy {
   createdOn: number;
   clonedBy?: string;
   clonedOn?: number;
-  originalRecipy?: string
+  originalRecipy?: string;
+  isSplitIntoGroups: IngredientsGroup[];
+  isBaseRecipy: boolean;
+  notApproved?: boolean;
+  source: string;
+  isCheckedAndApproved?: boolean;
+}
+
+export class emptyRecipy implements NewRecipy {
+  name: string = '';
+  ingrediends: Ingredient[] = [];
+  complexity: Complexity = Complexity.simple;
+  steps: PreparationStep[] = [];
+  type: DishType[] = [];
+  author: string;
+  createdOn: number = Date.now();
+  isSplitIntoGroups: IngredientsGroup[] = [];
+  isBaseRecipy: boolean = false;
+  source: string = '';
+  constructor(author: string) {
+    this.author = author;
+  }
 }
