@@ -1,133 +1,48 @@
+import { ProductType } from './../models/products.interface';
 import { Injectable } from '@angular/core';
+import { Store } from '@ngrx/store';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { take } from 'rxjs/operators';
+import { IAppState } from 'src/app/store/reducers';
 
+import * as fromRecipiesActions from '../../store/actions/recipies.actions';
 import { MeasuringUnit } from '../models/measuring-units.enum';
 import { Product } from '../models/products.interface';
 import { UserService } from './../../auth/services/user.service';
 import { Ingredient } from './../models/ingredient.interface';
 import { NewRecipy, Recipy } from './../models/recipy.interface';
-import { AmountConverterService } from './amount-converter.service';
 import { ProductsApiService } from './products-api.service';
 import { RecipiesApiService } from './recipies-api.service';
+import {
+  getAmountInSelectedUnit,
+  getCalorificValue,
+  getDefaultMeasuringUnit,
+  getIngredientIdFromName,
+  getIngredientText,
+  getProductText,
+  isIngrIncludedInAmountCalculation,
+  transformToGr,
+} from './recipies.utils';
 
 @Injectable({
   providedIn: 'root',
 })
 export class RecipiesService {
-  allRecipies$ = new BehaviorSubject<Recipy[]>([]);
   products$ = new BehaviorSubject<Product[]>([]);
   productsUpdated$ = new Subject<any>();
-  userRecipiesUpdated$ = new Subject();
-  recipiesUpdated$ = new Subject();
+  productsLoaded$ = new BehaviorSubject<boolean>(false);
+
+  filteredRecipies: number = 0;
 
   constructor(
     private recipiesApi: RecipiesApiService,
-    private converter: AmountConverterService,
     private productsApi: ProductsApiService,
-    private userService: UserService
+    private userService: UserService,
+    private store: Store<IAppState>
   ) {}
 
-  addRecipy(recipy: NewRecipy){
-    this.recipiesApi.addRecipy(recipy).subscribe((id: any) => {
-      this.addRecipyToUserRecipies(id.name);
-      this.recipiesUpdated$.next();
-    });
-  }
-
-  updateRecipy(recipy: Recipy){
-    this.recipiesApi.updateRecipy(recipy.id, recipy).pipe(take(1))
-    .subscribe(() => {
-      this.recipiesUpdated$.next();
-    });
-  }
-
-  getIngredientIdFromName(ingr: any){
-    let productId;
-    for (let product of this.products$.value) {
-      if (product.name === ingr.ingredient) {
-        productId = product.id;
-      }
-    }
-    return productId;
-  }
-
-  transformToGr(ingr: Ingredient) {
-    let unit: MeasuringUnit = ingr.defaultUnit;
-    let amount: number = ingr.amount;
-    let density: number = this.getDensity(ingr.product);
-    let grInOneItem: number = this.getGrPerItem(ingr.product);
-    switch (unit) {
-      case MeasuringUnit.gr:
-        return amount;
-        break;
-      case MeasuringUnit.kg:
-        return this.converter.kgToGR(amount);
-        break;
-      case MeasuringUnit.bunch:
-        return this.converter.bunchToGr(amount);
-        break;
-      case MeasuringUnit.coffeeSpoon:
-        return this.converter.coffeeSpoonsToGr(amount, density);
-        break;
-      case MeasuringUnit.dessertSpoon:
-        return this.converter.dessertSpoonsToGr(amount, density);
-        break;
-      case MeasuringUnit.l:
-        return this.converter.literToGr(amount, density);
-        break;
-      case MeasuringUnit.ml:
-        return this.converter.mlToGr(amount, density);
-        break;
-      case MeasuringUnit.pinch:
-        return this.converter.pinchToGr(amount);
-        break;
-      case MeasuringUnit.tableSpoon:
-        return this.converter.tableSpoonsToGr(amount, density);
-        break;
-      case MeasuringUnit.teaSpoon:
-        return this.converter.teaSpoonsToGr(amount, density);
-      case MeasuringUnit.item:
-        return this.converter.itemsToGr(amount, grInOneItem);
-    }
-  }
-
-  getDensity(productId: string) {
-    let density = 0;
-    for (let item of this.products$.value) {
-      if (item.id === productId) {
-        density = item.density;
-      }
-    }
-    return density;
-  }
-
-  getGrPerItem(productId: string) {
-    let grInOneItem = 0;
-    for (let item of this.products$.value) {
-      if (item.id === productId) {
-        grInOneItem = item.grInOneItem ? item.grInOneItem : 0;
-      }
-    }
-    return grInOneItem;
-  }
-
-  getRecipies() {
-    this.recipiesApi
-      .getRecipies()
-      .pipe(take(1))
-      .subscribe((res) => {
-        let array = Object.entries(res);
-        let recipies: any = [];
-        for (let entry of array) {
-          let recipy: any = {
-            id: entry[0],
-            ...entry[1],
-          };
-          recipies.push(recipy);
-        }
-        this.allRecipies$.next(recipies);
-      });
+  getIngredientIdFromName(ingr: any): string {
+    return getIngredientIdFromName(ingr, this.products$.value);
   }
 
   getRecipyById(id: string) {
@@ -154,6 +69,7 @@ export class RecipiesService {
           };
           products.push(recipy);
         }
+        this.productsLoaded$.next(true)
         this.products$.next(products);
       });
   }
@@ -166,105 +82,37 @@ export class RecipiesService {
   }
 
   getIngredientText(ingr: Ingredient): string {
-    let productName = '';
-    for (let product of this.products$.value) {
-      if (product.id === ingr.product) {
-        productName = product.name;
-      }
-    }
-    return productName;
+    return getIngredientText(ingr, this.products$.value);
   }
 
-  convertAmountToSelectedUnit(unit: MeasuringUnit, ingredient: Ingredient) {
-    if (unit == MeasuringUnit.gr) {
-      return ingredient.amount;
-    } else {
-      let amount = 0;
-      switch (unit) {
-        case MeasuringUnit.kg:
-          amount = this.converter.grToKg(ingredient.amount);
-          break;
-        case MeasuringUnit.l:
-          amount = this.converter.grToLiter(
-            ingredient.amount,
-            this.getDensity(ingredient.product)
-          );
-          break;
-        case MeasuringUnit.ml:
-          amount = this.converter.grToMl(
-            ingredient.amount,
-            this.getDensity(ingredient.product)
-          );
-          break;
-        case MeasuringUnit.tableSpoon:
-          amount = this.converter.grToTableSpoons(
-            ingredient.amount,
-            this.getDensity(ingredient.product)
-          );
-          break;
-        case MeasuringUnit.dessertSpoon:
-          amount = this.converter.grToDessertSpoons(
-            ingredient.amount,
-            this.getDensity(ingredient.product)
-          );
-          break;
-        case MeasuringUnit.teaSpoon:
-          amount = this.converter.grToTeaSpoons(
-            ingredient.amount,
-            this.getDensity(ingredient.product)
-          );
-          break;
-        case MeasuringUnit.coffeeSpoon:
-          amount = this.converter.grToCoffeeSpoons(
-            ingredient.amount,
-            this.getDensity(ingredient.product)
-          );
-          break;
-        case MeasuringUnit.pinch:
-          amount = this.converter.grToPinch(ingredient.amount);
-          break;
-        case MeasuringUnit.bunch:
-          amount = this.converter.grToBunch(ingredient.amount);
-          break;
-        case MeasuringUnit.item:
-          amount = this.converter.grToItems(
-            ingredient.amount,
-            this.getGrPerItem(ingredient.product)
-          );
-      }
-      return amount;
-    }
+  getProductNameById(id: string): string {
+    return getProductText(id, this.products$.value);
   }
 
-  addRecipyToUserRecipies(recipyId: string) {
-    let recipies = [];
-    if (this.userService.currentUser.recipies) {
-      recipies = this.userService.currentUser.recipies;
-    }
-    recipies.push(recipyId);
-    let updatedUserRecipies = { recipies: recipies };
-    this.userService
-      .updateUserDetailsFromMyDatabase(updatedUserRecipies)
-      .pipe(take(1))
-      .subscribe((res) => this.getRecipies());
+  getDefaultMU(id: string): MeasuringUnit{
+    return getDefaultMeasuringUnit(id, this.products$.value)
   }
 
-  removeRecipyFromUserRecipies(recipyId: string) {
-    let recipies = this.userService.currentUser.recipies;
-    let index = recipies.indexOf(recipyId);
-    recipies.splice(index, 1);
-    let updatedUserRecipies = { recipies: recipies };
-    this.userService
-      .updateUserDetailsFromMyDatabase(updatedUserRecipies)
-      .pipe(take(1))
-      .subscribe((res) => this.getRecipies());
+  getAmountInSelectedUnit(
+    selectedUnit: MeasuringUnit,
+    ingredientId: string,
+    amountInGr: number
+  ){
+    return getAmountInSelectedUnit(selectedUnit, ingredientId, amountInGr, this.products$.value)
+  }
+
+  getProductType(ingredientId: string){
+    let product = this.products$.value.find(product => product.id == ingredientId);
+    if(product){
+      return product.type
+    } else return ProductType.spice
   }
 
   deleteRecipy(recipy: Recipy) {
     this.recipiesApi
       .deleteRecipy(recipy.id)
       .pipe(take(1))
-      .subscribe((res) => this.recipiesUpdated$.next());
+      .subscribe((res) => {});
   }
 
   getRecipyBelongsTo(recipy: Recipy) {
@@ -281,9 +129,104 @@ export class RecipiesService {
     } else return recipy.createdOn;
   }
 
-  getIsUserRecipy(recipy: Recipy){
-    if(this.userService.currentUser.recipies){
-      return this.userService.currentUser.recipies.includes(recipy.id)
-    } else return false
+  saveUpdatedRecipy(recipy: Recipy, currentUserEmail: string) {
+    if (
+      currentUserEmail == recipy.author ||
+      ('clonedBy' in recipy && recipy.clonedBy == currentUserEmail)
+    ) {
+      let updatedRecipy: Recipy = {
+        id: recipy.id,
+        name: recipy.name,
+        ingrediends: recipy.ingrediends,
+        complexity: recipy.complexity,
+        steps: recipy.steps,
+        type: recipy.type,
+        isSplitIntoGroups: recipy.isSplitIntoGroups,
+        author: recipy.author,
+        createdOn: recipy.createdOn,
+        editedBy: currentUserEmail,
+        lastEdited: Date.now(),
+        isBaseRecipy: recipy.isBaseRecipy,
+        source: recipy.source
+      };
+      if ('clonedBy' in recipy) {
+        (updatedRecipy.clonedBy = recipy.clonedBy),
+          (updatedRecipy.clonedOn = recipy.clonedOn);
+      }
+      if ('photo' in recipy) {
+        updatedRecipy.photo = recipy.photo;
+      }
+      updatedRecipy.notApproved = this.checkIsRecipyApproved(updatedRecipy);
+      this.store.dispatch(
+        new fromRecipiesActions.UpdateRecipyAction(updatedRecipy)
+      );
+    } else {
+      let clonedRecipy: NewRecipy = {
+        name: recipy.name,
+        ingrediends: recipy.ingrediends,
+        complexity: recipy.complexity,
+        steps: recipy.steps,
+        type: recipy.type,
+        isSplitIntoGroups: recipy.isSplitIntoGroups,
+        author: recipy.author,
+        createdOn: recipy.createdOn,
+        clonedBy: currentUserEmail,
+        clonedOn: Date.now(),
+        isBaseRecipy: recipy.isBaseRecipy,
+        source: recipy.source
+      };
+      if ('photo' in recipy) {
+        clonedRecipy.photo = recipy.photo;
+      }
+      clonedRecipy.notApproved = this.checkIsRecipyApproved(clonedRecipy);
+      this.store.dispatch(
+        new fromRecipiesActions.AddNewRecipyAction(clonedRecipy)
+      );
+    }
+  }
+
+  checkIsRecipyApproved(recipy: Recipy | NewRecipy): boolean {
+    return !recipy.ingrediends.find((ingr) =>
+      this.getIsIngredientInDB(ingr.product)
+    );
+  }
+
+  transformToGr(ingr: Ingredient, amount: number) {
+    return transformToGr(ingr, amount, ingr.defaultUnit, this.products$.value);
+  }
+
+  getIsIngredientInDB(id: string) {
+    return this.products$.value.find((ingr) => ingr.id == id);
+  }
+
+  getIsIngredientIncludedInAmountCalculation(ingr: Ingredient): boolean {
+    return isIngrIncludedInAmountCalculation(ingr, this.products$.value);
+  }
+
+  getCoeficient(
+    ingredients: Ingredient[],
+    portionsToServe: number,
+    portionSize: number
+  ) {
+    let amount = 0;
+    for (let ingr of ingredients) {
+      if (
+        this.getIsIngredientInDB(ingr.product) &&
+        this.getIsIngredientIncludedInAmountCalculation(ingr)
+      ) {
+        amount = ingr.amount + amount; // amount of ingreds with calories
+      }
+    }
+    return (portionsToServe * portionSize) / amount;
+  }
+
+  countRecipyCalorificValue(ingreds: Ingredient[]){
+    let calories = 0;
+    let totalAmount = 0;
+    ingreds.forEach(ingr => {
+      totalAmount += ingr.amount;
+      calories += ingr.amount * getCalorificValue(ingr, this.products$.value)
+    })
+    return calories / totalAmount
   }
 }
